big fix

+++ src/main/java/com/adventurequest/util/DefaultGamesCountUtil.java
@@ -1,12 +1,13 @@
 package com.adventurequest.util;
 
 public final class DefaultGamesCountUtil {
-    private static final int DEFAULT_GAME_COUNT = 0;
+    private static final Long DEFAULT_GAME_COUNT = 0L;
 
     private DefaultGamesCountUtil() {
     }
 
-    public static int getDefaultGamesCount() {
+    public static Long getDefaultGamesCount() {
         return DEFAULT_GAME_COUNT;
     }
 }
+
